requirements.txt
asgiref==3.2.3
boto3==1.11.12
botocore==1.14.12
certifi==2019.11.28
chardet==3.0.4
defusedxml==0.6.0
Django==3.0.2
django-allauth==0.41.0
django-rest-auth==0.9.5
django-storages==1.9.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
docutils==0.15.2
idna==2.8
jmespath==0.9.4
mysqlclient==1.4.6
oauthlib==3.1.0
Pillow==7.0.0
PyJWT==1.7.1
python-dateutil==2.8.1
python3-openid==3.1.0
pytz==2019.3
requests==2.22.0
requests-oauthlib==1.3.0
s3transfer==0.3.3
six==1.13.0
sqlparse==0.3.0
urllib3==1.25.7
